# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Livraria Store"
  },
  "host": "localhost:10010",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/books": {
      "x-swagger-router-controller": "books",
      "get": {
        "tags":[
          "Book"
        ],
        "description": "Return books of the store",
        "operationId": "getBooks",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for books can be title, author, category or publisher",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Actual page",
            "required": false,
            "type": "string"
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      },
      "post":{
        "tags":[
          "Book"
        ],
        "description":"Create a book on the store",
        "operationId": "addBook",
        "parameters":[
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema":{
              "$ref": "#/definitions/BookPost"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "400":{
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/v1/books/{bookId}": {
      "x-swagger-router-controller": "books",
      "get": {
        "tags":[
          "Book"
        ],
        "description": "Return a book",
        "operationId": "getBook",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "Id of the book to return",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "404":{
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put":{
        "tags":[
          "Book"
        ],
        "description":"Edit a book on the store",
        "operationId": "editBook",
        "parameters":[
          {
            "name": "bookId",
            "in": "path",
            "description": "Id of the book to return",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema":{
              "$ref": "#/definitions/BookPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Altered with success",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "400":{
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete":{
        "tags":[
          "Book"
        ],
        "description":"Delete a book on the store",
        "operationId": "deleteBook",
        "parameters":[
          {
            "name": "bookId",
            "in": "path",
            "description": "Id of the book to delete",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "schema": {
              "$ref": "#/definitions/DeleteResponse"
            }
          },
          "400":{
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/v1/books/{bookId}/comments": {
      "x-swagger-router-controller": "books",
      "get": {
        "tags":[
          "Book"
        ],
        "description": "Return the comments of a book",
        "operationId": "getBookComments",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "Id of the book to return the comments",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Actual page",
            "required": false,
            "type": "string"
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Comments"
            }
          },
          "404":{
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put":{
        "tags":[
          "Book"
        ],
        "description":"Add a comment to this book",
        "operationId": "editBookComment",
        "parameters":[
          {
            "name": "bookId",
            "in": "path",
            "description": "Id of the book to commnet",
            "required": false,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Body",
            "required": true,
            "schema":{
              "$ref": "#/definitions/BookPost"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Altered with success",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "400":{
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AuthorWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "name":{
          "type":"string"
        },
        "author_uri":{
          "type":"string"
        }
      }
    },
...
    "PubCompanyWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "name":{
         "type":"string" 
        },
        "pub_company_uri":{
          "type":"string"
        }
      }
    },
...
    "CategoryWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "name":{
         "type":"string" 
        },
        "catgory_uri":{
          "type":"string" 
         }
      }
    },
    "BookPost":{
      "required": [
        "name",
        "author_id",
        "pub_company_id",
        "category_id",
        "price",
        "year"
      ],
      "properties":{
        "name":{
          "type":"string"
        },
        "author_id":{
          "type":"number"
        },
        "pub_company_id":{
          "type":"number"
        },
        "category_id":{
          "type":"number"
        },
        "price":{
          "type":"number"
        },
        "year":{
          "type":"number"
        }
      }
    },
    "Book":{
      "properties":{
        "id":{
          "type":"number"
        },
        "name":{
          "type":"number"
        },
        "author":{
          "$ref":"#/definitions/AuthorWithUri"
        },
        "pub_company":{
          "$ref":"#/definitions/PubCompanyWithUri"
        },
        "category":{
          "$ref":"#/definitions/CategoryWithUri"
        },
        "price":{
          "type":"number"
        },
        "year":{
          "type":"number"
        }
      }
    },
    "BookWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "author":{
          "$ref":"#/definitions/AuthorWithUri"
        },
        "pub_company":{
          "$ref":"#/definitions/PubCompanyWithUri"
        },
        "category":{
          "$ref":"#/definitions/CategoryWithUri"
        },
        "name":{
          "type":"number"
        },
        "price":{
          "type":"number"
        },
        "year":{
          "type":"string"
        },
        "livro_uri":{
          "type":"string"
        }
      }
    },
...
    "UserWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "name":{
          "type":"string"
        },
        "cpf":{
          "type":"string"
        },
        "user_uri":{
          "type":"string"
        }
      }
    },
...
    "CommentWithUri":{
      "properties":{
        "id":{
          "type":"number"
        },
        "comment":{
          "type":"string"
        },
        "stars":{
          "type":"number"
        },
        "user":{
          "$ref":"#/definitions/UserWithUri"
        },
        "book":{
          "$ref":"#/definitions/BookWithUri"
        },
        "comment_uri":{
          "type":"string"
        }
      }
    },
...
    "Comments": {
      "required": [],
      "properties": {
        "pages": {
            "type": "number"
        },
        "actualPage": {
            "type": "number"
        },
        "nextPage": {
            "type": "string"
        },
        "data": {
          "type": "array",
          "xml": {
            "name": "data",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/CommentWithUri"
          }
        }
      }
    },
...
    "Books": {
      "required": [],
      "properties": {
        "pages": {
            "type": "number"
        },
        "actualPage": {
            "type": "number"
        },
        "nextPage": {
            "type": "string"
        },
        "data": {
          "type": "array",
          "xml": {
            "name": "data",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/BookWithUri"
          }
        }
      }
    },
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "DeleteResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}